Fix error response, clean up a bit
Vasco was returning a 500 error on any bad server response, which was
by design but would actually cause the load balancer to shut down the
server. So now if vasco is awake, it returns 200 to a basic status
ping, and detailed server requests return more elaborate return codes.

Also, there was a minor bug in responding to registration requests.