Report status for expected services
We now take a list of expected services in the environment, and status
reporting includes a list of services expected but not found.

Until we have the ability to run multiple parallel copies of Vasco,
however, failure of expectations does not cause a bad status return (we
still return 200). This is because otherwise, AWS will shut down a
server if its status query returns non-200.